javadoc created and fixed repository

+++ src/main/java/com/nc/lab1/person/Person.java
@@ -9,20 +9,57 @@ import lombok.AllArgsConstructor;
 import java.time.LocalDate;
 import java.time.Period;
 
+/**
+ * Description of a person.
+ * @author Nikolay Evsyukov
+ * @version 1.0
+ */
 @Getter
 @Setter
 @AllArgsConstructor
 @ToString
 @EqualsAndHashCode()
 public class Person {
-    private long id;
+
+    /**
+     * Person ID.
+     */
+    private int id;
+
+    /**
+     * The last name of the person.
+     */
     private String surname;
+
+    /**
+     * The name of the person.
+     */
     private String name;
+
+    /**
+     * Patronymic of the person.
+     */
     private String patronymic;
+
+    /**
+     * Date of birth of the person.
+     */
     private LocalDate dateOfBirth;
+
+    /**
+     * Human gender.
+     */
     private Gender gender;
+
+    /**
+     * Person's passport.
+     */
     private Passport passport;
 
+    /**
+     * The age of the person.
+     * @return age
+     */
     public int getAge(){
         LocalDate currentDate = LocalDate.now();
         return Period.between(dateOfBirth,currentDate).getYears();
